+++ taskmonitor/helpers.py
@@ -42,12 +42,17 @@ def next_number(key: str = None) -> int:
     return next_number._counter[key].__next__()
 
 
+def dict_sort_keys(d: dict) -> dict:
+    """Return a copy of this dictionary with sorted keys."""
+    return dict(sorted(d.items(), key=lambda x: x[0].lower()))
+
+
 def truncate_args(args: list) -> list:
     """Truncate nested elements and return as new list.
 
-    Dicts will be replaced by `{"": None}`
+    Dicts will be replaced by `{}`
 
-    Lists, tuple and sets wil be replaced by `[None]`
+    Lists, tuple and sets wil be replaced by `[]`
     """
     return [_replace_nested_element(item) for item in args]
 
@@ -56,14 +61,14 @@ def truncate_kwargs(kwargs: dict) -> dict:
     """Truncate nested values and return as new dict.
 
     Example:
-    `{"a": {"aa": 1, ...}, "b": [1, 2, ...]}`-> `{"a": {"": None}, "b": [None]}`
+    `{"a": {"aa": 1, ...}, "b": [1, 2, ...]}`-> `{"a": {}, "b": []}`
     """
     return {key: _replace_nested_element(value) for key, value in kwargs.items()}
 
 
 def _replace_nested_element(value):
     if isinstance(value, dict):
-        return {"": None}
+        return dict()
     elif isinstance(value, (list, tuple, set)):
-        return [None]
+        return list()
     return value